--- .eslintcache
@@ -1,1 +1,0 @@
-[{"/Users/lucid/Desktop/mv-react/react-bookstore/src/components/BooksList.js":"1","/Users/lucid/Desktop/mv-react/react-bookstore/src/store.js":"2"},{"size":603,"mtime":1611916171147,"results":"3","hashOfConfig":"4"},{"size":950,"mtime":1611916477662,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"13qasxl",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/lucid/Desktop/mv-react/react-bookstore/src/components/BooksList.js",[],"/Users/lucid/Desktop/mv-react/react-bookstore/src/store.js",[]]
