--- src/App.js
@@ -1,14 +1,0 @@
-import './App.css';
-import BooksList from './components/BooksList';
-import BooksForm from './components/BooksForm';
-
-function App() {
-  return (
-    <div>
-      <BooksList />
-      <BooksForm />
-    </div>
-  );
-}
-
-export default App;
