+++ src/components/Book.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 
 const Book = props => {
   const { id, title, category } = props;
+
   return (
     <tr>
       <th>{id}</th>
